Reorganize code into api, meta, util modules
